+++ glsl/mirror.frag
@@ -4,7 +4,7 @@
 #endif
 layout(set=0,binding=0,std140) uniform World {
 	vec3 CAMERA_POSITION;
-	float ENVIRONMENT_MIPS;
+	uint ENVIRONMENT_MIPS;
 };
 
 layout(set=0, binding=1) uniform samplerCube ENVIRONMENT;
